# docs/swagger.yaml
basePath: /
definitions:
  github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: Success
        type: string
      meta:
        $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.PaginationMeta'
      status:
        example: success
        type: string
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.CreateProductRequest:
    properties:
      description:
        example: High performance gaming laptop
        type: string
      image:
        example: products/laptop-gaming.jpg
        type: string
      name:
        example: Laptop Gaming
        type: string
      price:
        example: 15000000
        type: number
      stock:
        example: 50
        minimum: 0
        type: integer
    required:
    - name
    - price
    - stock
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.PaginationMeta:
    properties:
      has_next:
        example: true
        type: boolean
      has_previous:
        example: false
        type: boolean
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.UpdateProductRequest:
    properties:
      description:
        example: Updated high performance gaming laptop
        type: string
      image:
        example: products/laptop-gaming-pro.jpg
        type: string
      name:
        example: Laptop Gaming Pro
        type: string
      price:
        example: 17000000
        type: number
      stock:
        example: 30
        type: integer
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@hrmis.com
    name: API Support
  description: HRMIS Backend API - Human Resource Management Information System
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: HRMIS API
  version: "1.0"
paths:
...
  /api/v1/products:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all products sorted by newest first
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product with the provided data. If image is from temp
        folder, it will be moved to permanent folder.
      parameters:
      - description: Product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Products
  /api/v1/products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product by its UUID
      parameters:
      - description: Product UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Retrieve a single product by its UUID
      parameters:
      - description: Product UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      summary: Get product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update an existing product by its UUID
      parameters:
      - description: Product UUID
        in: path
        name: id
        required: true
        type: string
      - description: Product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - Products
  /api/v1/products/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image file to S3/MinIO storage (temporary folder)
      parameters:
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
            - properties:
                data:
                  properties:
                    image:
                      type: string
                    preview_expires_in:
                      type: integer
                    preview_url:
                      type: string
                    upload_expires_in:
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Upload product image
      tags:
      - Products
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"